make preview work
